refactor(backend): group tryout score fixture
